+++ randovania/generator/hint_distributor.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import dataclasses
 import functools
+import logging
 import math
 from abc import ABC, abstractmethod
 from collections import defaultdict
@@ -27,7 +28,8 @@ from randovania.game_description.hint_features import HintFeature
 from randovania.game_description.resources.pickup_index import PickupIndex
 from randovania.generator.filler.filler_library import UnableToGenerate
 from randovania.generator.filler.player_state import HintState, PlayerState
-from randovania.resolver import debug
+from randovania.layout.base.dock_rando_configuration import DockRandoMode
+from randovania.resolver import debug, resolver
 
 if TYPE_CHECKING:
     from randovania.game_description.assignment import PickupTarget
@@ -36,6 +38,8 @@ if TYPE_CHECKING:
     from randovania.game_description.pickup.pickup_entry import PickupEntry
     from randovania.generator.filler.filler_configuration import FillerResults, PlayerPool
     from randovania.generator.pre_fill_params import PreFillParams
+    from randovania.layout.base.base_configuration import BaseConfiguration
+    from randovania.resolver.hint_state import ResolverHintState
 
 HintProvider = Callable[[PlayerState, GamePatches, Random, PickupIndex], LocationHint | None]
 
@@ -43,6 +47,9 @@ HintTargetPrecision = tuple[PickupIndex, PrecisionPair]
 
 HintFeatureGaussianParams = tuple[float, float]
 """The mean and standard deviation defining a Gaussian distribution."""
+
+
+logger = logging.getLogger(__name__)
 
 
 class FeatureChooser[PrecisionT: Enum]:
@@ -256,15 +263,16 @@ class HintDistributor(ABC):
         patches: GamePatches,
         rng: Random,
         player_pool: PlayerPool,
-        player_state: PlayerState,
-        player_pools: list[PlayerPool],
+        region_list: RegionList,
+        hint_state: HintState,
+        player_pools: Sequence[PlayerPool],
     ) -> GamePatches:
         # Since we haven't added expansions yet, these hints will always be for items added by the filler.
         full_hints_patches = self.fill_unassigned_hints(
             patches,
-            player_state.game.region_list,
+            region_list,
             rng,
-            player_state.hint_state,
+            hint_state,
             player_pools,
         )
         return await self.assign_precision_to_hints(full_hints_patches, rng, player_pool, player_pools)
@@ -274,7 +282,7 @@ class HintDistributor(ABC):
         patches: GamePatches,
         rng: Random,
         player_pool: PlayerPool,
-        player_pools: list[PlayerPool],
+        player_pools: Sequence[PlayerPool],
         hint_kinds: Container[HintNodeKind] = {HintNodeKind.GENERIC},
     ) -> GamePatches:
         """
@@ -289,7 +297,7 @@ class HintDistributor(ABC):
 
     @final
     @staticmethod
-    def hint_suitability_for_target(target: PickupTarget, player_pools: list[PlayerPool]) -> HintSuitability:
+    def hint_suitability_for_target(target: PickupTarget, player_pools: Sequence[PlayerPool]) -> HintSuitability:
         """
         Determines the HintSuitability for the given target,
         according to the criteria of its *owner's* HintDistributor.
@@ -319,7 +327,7 @@ class HintDistributor(ABC):
         region_list: RegionList,
         rng: Random,
         hint_state: HintState,
-        player_pools: list[PlayerPool],
+        player_pools: Sequence[PlayerPool],
     ) -> GamePatches:
         """
         Selects targets for all remaining unassigned generic hint nodes
@@ -344,14 +352,28 @@ class HintDistributor(ABC):
             real_potential_targets = {
                 target
                 for target in sorted(potential_targets)
-                if self.hint_suitability_for_target(patches.pickup_assignment[target], player_pools)
+                if target in patches.pickup_assignment
+                and self.hint_suitability_for_target(patches.pickup_assignment[target], player_pools)
                 >= HintSuitability.MORE_INTERESTING
             }
             # don't hint things twice
             real_potential_targets -= hinted_locations
+            debug.debug_print(f"  * Trying {len(real_potential_targets)} logical interesting pickups")
 
             if not real_potential_targets:
-                # no interesting pickups to place - use anything placed during fill or prefill
+                # no logical pickups to place - use any interesting pickups
+                real_potential_targets = {
+                    pickup
+                    for pickup, entry in patches.pickup_assignment.items()
+                    if self.hint_suitability_for_target(entry, player_pools) >= HintSuitability.MORE_INTERESTING
+                }
+                real_potential_targets -= hinted_locations
+                debug.debug_print(
+                    f"  * No logical interesting pickups; trying {len(real_potential_targets)} interesting pickups"
+                )
+
+            if not real_potential_targets:
+                # no interesting pickups to place - use less interesting pickups
                 real_potential_targets = {
                     pickup
                     for pickup, entry in patches.pickup_assignment.items()
@@ -651,20 +673,78 @@ class AllJokesHintDistributor(HintDistributor):
         patches: GamePatches,
         rng: Random,
         player_pool: PlayerPool,
-        player_pools: list[PlayerPool],
+        player_pools: Sequence[PlayerPool],
         hint_kinds: Container[HintNodeKind] = {HintNodeKind.GENERIC},
     ) -> GamePatches:
         return self.replace_hints_without_precision_with_jokes(patches)
 
 
-async def distribute_specific_location_hints(
-    rng: Random, filler_results: FillerResults, player_pools: list[PlayerPool]
+def _should_use_resolver_hints(config: BaseConfiguration) -> bool:
+    return config.use_resolver_hints or config.dock_rando.mode == DockRandoMode.DOCKS
+
+
+async def get_resolver_hint_state(player: int, patches: GamePatches) -> ResolverHintState | None:
+    with debug.with_level(0):
+        new_state = await resolver.resolve(patches.configuration, patches, collect_hint_data=True)
+
+    if new_state is None:
+        logger.warning(
+            f"Unable to solve game for player {player + 1} after placing doors. Hints will be based on generator state."
+        )
+        return None
+    else:
+        debug.debug_print(f">> Player {player + 1} is solve-able after door placement. Beginning hint placement.")
+        return new_state.hint_state
+
+
+async def distribute_generic_hints(
+    rng: Random,
+    filler_results: FillerResults,
 ) -> FillerResults:
-    """Distribute HintNodeKind.SPECIFIC_PICKUP hints *after* all items have been placed."""
+    """Distribute HintNodeKind.GENERIC hints after all pickups are placed."""
+    new_patches: dict[int, GamePatches] = {
+        player: result.patches for player, result in filler_results.player_results.items()
+    }
+
+    for player_index, result in filler_results.player_results.items():
+        patches = result.patches
+
+        hint_state: HintState = result.hint_state
+        if _should_use_resolver_hints(patches.configuration):
+            resolver_hint_state = await get_resolver_hint_state(player_index, patches)
+            if resolver_hint_state is not None:
+                hint_state = resolver_hint_state
+
+        hint_distributor = patches.game.game.generator.hint_distributor
+        new_patches[player_index] = await hint_distributor.assign_post_filler_hints(
+            patches,
+            rng,
+            result.pool,
+            patches.game.region_list,
+            hint_state,
+            filler_results.player_pools,
+        )
+
+    return dataclasses.replace(
+        filler_results,
+        player_results={
+            player: dataclasses.replace(result, patches=new_patches[player])
+            for player, result in filler_results.player_results.items()
+        },
+    )
+
+
+async def distribute_specific_location_hints(
+    rng: Random,
+    filler_results: FillerResults,
+) -> FillerResults:
+    """Distribute HintNodeKind.SPECIFIC_PICKUP hints after all pickups are placed."""
     old_patches: dict[int, GamePatches] = {
         player: result.patches for player, result in filler_results.player_results.items()
     }
     new_patches: dict[int, GamePatches] = {}
+
+    player_pools = filler_results.player_pools
 
     for player_index, patches in old_patches.items():
         player_pool = player_pools[player_index]